requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Drive
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI/ML - Face Detection
insightface==0.7.3
onnxruntime-gpu==1.17.0
# onnxruntime==1.17.0  # Use this instead if no GPU

# AI/ML - Vision & Embeddings
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2
sentence-transformers==2.3.1
Pillow==10.2.0

# AI/ML - Audio Transcription
openai-whisper==20231117
faster-whisper==1.0.0

# Image Processing
opencv-python==4.9.0.80
imagehash==4.3.1

# Video Processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.1
numpy==1.26.3
pandas==2.1.4

# Clustering
scikit-learn==1.4.0
hdbscan==0.8.33

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# Static Site Generation
jinja2==3.1.3
markdown==3.5.2

# Logging
loguru==0.7.2

# HTTP Client
httpx==0.26.0
